backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.27.0
pydantic==2.13.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
openai==1.35.7
langchain==0.1.16
python-jose[cryptography]==3.3.0
slowapi==0.1.8
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
